# Rewrite Kernel#tap with Ruby