@shyouhei 2020-07-03
Second thought you might also want to do this for `Kernel#then`.