@benoittgt 2020-07-03
Related explanation submitted by @k0kubun [on twitter](https://twitter.com/k0kubun/status/1278954301834944512):
> I thought a block must be inlinable over methods which are currently written in C with future JIT, but it's actually hard to update (for inlining block) a part of C code from JIT. My current plan is to mix C and Ruby for methods like #each, but then I noticed #tap was too easy :)

😊