@k0kubun 2020-07-03
> Second thought you might also want to do this for Kernel#then.

Absolutely :slightly_smiling_face: I'll prepare another PR to see CI results for it separately though.