@ruurd 2020-07-10
> because Ruby is faster than C.

Lies. Damned lies. Statistics. Did you find out WHY it is faster?