@simi 2020-07-10
> Lies. Damned lies. Statistics. Did you find out WHY it is faster?

It is not meant it is faster in general, but in this case of `Kernel#tap` implementation. Stats are in original post.