@schneems 2020-11-19
> Everything is written in this thread actually.

I don't totally follow from this thread alone. However via the linked twitter conversation: https://twitter.com/k0kubun/status/1279128061355495425

> I haven't confirmed the details, but my current guess is that creating a new VM frame (vm_exec_core) is expensive and reusing the same VM frame by yield is faster than creating one every time like vm_exec_core -> rb_obj_tap -> vm_exec_core.

So to be explicit. While this unlocks JIT improvements, it is ALSO faster even without JIT. Possibly due to the reduction in calling vm_exec_core